render: Separate tesselation code into render/common_tess
