+++ render/common_tess/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "ruffle_render_common_tess"
+version = "0.1.0"
+authors = ["Mike Welsh <mwelsh@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+log = "0.4"
+lyon = "0.15.6"
+
+[dependencies.ruffle_core]
+path = "../../core"
+default-features = false
