+++ render/glium/src/lib.rs
@@ -3,19 +3,11 @@
 use glium::uniforms::{Sampler, UniformValue, Uniforms};
 use glium::{draw_parameters::DrawParameters, implement_vertex, uniform, Display, Frame, Surface};
 use glutin::WindowedContext;
-use lyon::path::Path;
-use lyon::tessellation::{
-    self,
-    geometry_builder::{BuffersBuilder, FillVertexConstructor, VertexBuffers},
-    FillAttributes, FillTessellator, StrokeAttributes, StrokeTessellator, StrokeVertexConstructor,
-};
 use ruffle_core::backend::render::swf::{self, FillStyle};
 use ruffle_core::backend::render::{
     BitmapHandle, BitmapInfo, Color, Letterbox, RenderBackend, ShapeHandle, Transform,
 };
-use ruffle_core::shape_utils::{DrawCommand, DrawPath};
 use std::convert::TryInto;
-use swf::Twips;
 
 type Error = Box<dyn std::error::Error>;
 
@@ -166,332 +158,83 @@ impl GliumRenderBackend {
     }
 
     fn register_shape_internal(&mut self, shape: &swf::Shape) -> ShapeHandle {
+        use ruffle_render_common_tess::DrawType as TessDrawType;
+
         let handle = ShapeHandle(self.meshes.len());
-        let paths = ruffle_core::shape_utils::swf_shape_to_paths(shape);
 
-        use lyon::tessellation::{FillOptions, StrokeOptions};
+        let lyon_mesh = ruffle_render_common_tess::tessellate_shape(shape, |id| {
+            self.textures
+                .iter()
+                .find(|(other_id, _tex)| *other_id == id)
+                .map(|tex| (tex.1.width, tex.1.height))
+        });
 
-        let mut mesh = Mesh { draws: vec![] };
+        let mut draws = Vec::with_capacity(lyon_mesh.len());
 
-        let mut fill_tess = FillTessellator::new();
-        let mut stroke_tess = StrokeTessellator::new();
-        let mut lyon_mesh: VertexBuffers<_, u32> = VertexBuffers::new();
-
-        fn flush_draw(
-            draw: DrawType,
-            mesh: &mut Mesh,
-            lyon_mesh: &mut VertexBuffers<Vertex, u32>,
-            display: &Display,
-        ) {
-            if lyon_mesh.vertices.is_empty() {
-                return;
-            }
-
-            let vertex_buffer = glium::VertexBuffer::new(display, &lyon_mesh.vertices[..]).unwrap();
-
-            let index_buffer = glium::IndexBuffer::new(
-                display,
-                glium::index::PrimitiveType::TrianglesList,
-                &lyon_mesh.indices[..],
+        for draw in lyon_mesh {
+            let vertex_buffer = glium::VertexBuffer::new(
+                &self.display,
+                &draw
+                    .vertices
+                    .into_iter()
+                    .map(|v| Vertex {
+                        position: v.position,
+                        color: v.color,
+                    })
+                    .collect::<Vec<_>>(),
             )
             .unwrap();
 
-            mesh.draws.push(Draw {
-                draw_type: draw,
-                vertex_buffer,
-                index_buffer,
-            });
+            let index_buffer = glium::IndexBuffer::new(
+                &self.display,
+                glium::index::PrimitiveType::TrianglesList,
+                &draw.indices,
+            )
+            .unwrap();
 
-            *lyon_mesh = VertexBuffers::new();
-        }
-
-        for path in paths {
-            match path {
-                DrawPath::Fill { style, commands } => match style {
-                    FillStyle::Color(color) => {
-                        let color = [
-                            f32::from(color.r) / 255.0,
-                            f32::from(color.g) / 255.0,
-                            f32::from(color.b) / 255.0,
-                            f32::from(color.a) / 255.0,
-                        ];
-
-                        let mut buffers_builder =
-                            BuffersBuilder::new(&mut lyon_mesh, RuffleVertexCtor { color });
-
-                        if let Err(e) = fill_tess.tessellate_path(
-                            &ruffle_path_to_lyon_path(commands, true),
-                            &FillOptions::even_odd(),
-                            &mut buffers_builder,
-                        ) {
-                            // This may just be a degenerate path; skip it.
-                            log::error!("Tessellation failure: {:?}", e);
-                            continue;
-                        }
-                    }
-                    FillStyle::LinearGradient(gradient) => {
-                        flush_draw(DrawType::Color, &mut mesh, &mut lyon_mesh, &self.display);
-
-                        let mut buffers_builder = BuffersBuilder::new(
-                            &mut lyon_mesh,
-                            RuffleVertexCtor {
-                                color: [1.0, 1.0, 1.0, 1.0],
-                            },
-                        );
-
-                        if let Err(e) = fill_tess.tessellate_path(
-                            &ruffle_path_to_lyon_path(commands, true),
-                            &FillOptions::even_odd(),
-                            &mut buffers_builder,
-                        ) {
-                            // This may just be a degenerate path; skip it.
-                            log::error!("Tessellation failure: {:?}", e);
-                            continue;
-                        }
-
-                        let mut colors: Vec<[f32; 4]> = Vec::with_capacity(8);
-                        let mut ratios: Vec<f32> = Vec::with_capacity(8);
-                        for record in &gradient.records {
-                            colors.push([
-                                f32::from(record.color.r) / 255.0,
-                                f32::from(record.color.g) / 255.0,
-                                f32::from(record.color.b) / 255.0,
-                                f32::from(record.color.a) / 255.0,
-                            ]);
-                            ratios.push(f32::from(record.ratio) / 255.0);
-                        }
-
-                        let uniforms = GradientUniforms {
-                            gradient_type: 0,
-                            ratios,
-                            colors,
-                            num_colors: gradient.records.len() as u32,
-                            matrix: swf_to_gl_matrix(gradient.matrix.clone()),
-                            repeat_mode: 0,
-                            focal_point: 0.0,
-                        };
-
-                        flush_draw(
-                            DrawType::Gradient(uniforms),
-                            &mut mesh,
-                            &mut lyon_mesh,
-                            &self.display,
-                        );
-                    }
-                    FillStyle::RadialGradient(gradient) => {
-                        flush_draw(DrawType::Color, &mut mesh, &mut lyon_mesh, &self.display);
-
-                        let mut buffers_builder = BuffersBuilder::new(
-                            &mut lyon_mesh,
-                            RuffleVertexCtor {
-                                color: [1.0, 1.0, 1.0, 1.0],
-                            },
-                        );
-
-                        if let Err(e) = fill_tess.tessellate_path(
-                            &ruffle_path_to_lyon_path(commands, true),
-                            &FillOptions::even_odd(),
-                            &mut buffers_builder,
-                        ) {
-                            // This may just be a degenerate path; skip it.
-                            log::error!("Tessellation failure: {:?}", e);
-                            continue;
-                        }
-
-                        let mut colors: Vec<[f32; 4]> = Vec::with_capacity(8);
-                        let mut ratios: Vec<f32> = Vec::with_capacity(8);
-                        for record in &gradient.records {
-                            colors.push([
-                                f32::from(record.color.r) / 255.0,
-                                f32::from(record.color.g) / 255.0,
-                                f32::from(record.color.b) / 255.0,
-                                f32::from(record.color.a) / 255.0,
-                            ]);
-                            ratios.push(f32::from(record.ratio) / 255.0);
-                        }
-
-                        let uniforms = GradientUniforms {
-                            gradient_type: 1,
-                            ratios,
-                            colors,
-                            num_colors: gradient.records.len() as u32,
-                            matrix: swf_to_gl_matrix(gradient.matrix.clone()),
-                            repeat_mode: 0,
-                            focal_point: 0.0,
-                        };
-
-                        flush_draw(
-                            DrawType::Gradient(uniforms),
-                            &mut mesh,
-                            &mut lyon_mesh,
-                            &self.display,
-                        );
-                    }
-                    FillStyle::FocalGradient {
-                        gradient,
-                        focal_point,
-                    } => {
-                        flush_draw(DrawType::Color, &mut mesh, &mut lyon_mesh, &self.display);
-
-                        let mut buffers_builder = BuffersBuilder::new(
-                            &mut lyon_mesh,
-                            RuffleVertexCtor {
-                                color: [1.0, 1.0, 1.0, 1.0],
-                            },
-                        );
-
-                        if let Err(e) = fill_tess.tessellate_path(
-                            &ruffle_path_to_lyon_path(commands, true),
-                            &FillOptions::even_odd(),
-                            &mut buffers_builder,
-                        ) {
-                            // This may just be a degenerate path; skip it.
-                            log::error!("Tessellation failure: {:?}", e);
-                            continue;
-                        }
-
-                        let mut colors: Vec<[f32; 4]> = Vec::with_capacity(8);
-                        let mut ratios: Vec<f32> = Vec::with_capacity(8);
-                        for record in &gradient.records {
-                            colors.push([
-                                f32::from(record.color.r) / 255.0,
-                                f32::from(record.color.g) / 255.0,
-                                f32::from(record.color.b) / 255.0,
-                                f32::from(record.color.a) / 255.0,
-                            ]);
-                            ratios.push(f32::from(record.ratio) / 255.0);
-                        }
-
-                        let uniforms = GradientUniforms {
-                            gradient_type: 1,
-                            ratios,
-                            colors,
-                            num_colors: gradient.records.len() as u32,
-                            matrix: swf_to_gl_matrix(gradient.matrix.clone()),
-                            repeat_mode: 0,
-                            focal_point: *focal_point,
-                        };
-
-                        flush_draw(
-                            DrawType::Gradient(uniforms),
-                            &mut mesh,
-                            &mut lyon_mesh,
-                            &self.display,
-                        );
-                    }
-                    FillStyle::Bitmap {
-                        id,
-                        matrix,
-                        is_smoothed,
-                        is_repeating,
-                    } => {
-                        flush_draw(DrawType::Color, &mut mesh, &mut lyon_mesh, &self.display);
-
-                        let mut buffers_builder = BuffersBuilder::new(
-                            &mut lyon_mesh,
-                            RuffleVertexCtor {
-                                color: [1.0, 1.0, 1.0, 1.0],
-                            },
-                        );
-
-                        if let Err(e) = fill_tess.tessellate_path(
-                            &ruffle_path_to_lyon_path(commands, true),
-                            &FillOptions::even_odd(),
-                            &mut buffers_builder,
-                        ) {
-                            // This may just be a degenerate path; skip it.
-                            log::error!("Tessellation failure: {:?}", e);
-                            continue;
-                        }
-
-                        let texture = &self
-                            .textures
-                            .iter()
-                            .find(|(other_id, _tex)| *other_id == *id)
-                            .unwrap()
-                            .1;
-
-                        let uniforms = BitmapUniforms {
-                            matrix: swf_bitmap_to_gl_matrix(
-                                matrix.clone(),
-                                texture.width,
-                                texture.height,
-                            ),
-                            id: *id,
-                        };
-
-                        flush_draw(
-                            DrawType::Bitmap {
-                                uniforms,
-                                is_smoothed: *is_smoothed,
-                                is_repeating: *is_repeating,
-                            },
-                            &mut mesh,
-                            &mut lyon_mesh,
-                            &self.display,
-                        );
-                    }
+            let out_draw = match draw.draw_type {
+                TessDrawType::Color => Draw {
+                    draw_type: DrawType::Color,
+                    vertex_buffer,
+                    index_buffer,
                 },
-                DrawPath::Stroke {
-                    style,
-                    commands,
-                    is_closed,
-                } => {
-                    let color = [
-                        f32::from(style.color.r) / 255.0,
-                        f32::from(style.color.g) / 255.0,
-                        f32::from(style.color.b) / 255.0,
-                        f32::from(style.color.a) / 255.0,
-                    ];
-
-                    let mut buffers_builder =
-                        BuffersBuilder::new(&mut lyon_mesh, RuffleVertexCtor { color });
-
-                    // TODO(Herschel): 0 width indicates "hairline".
-                    let width = if style.width.to_pixels() >= 1.0 {
-                        style.width.to_pixels() as f32
-                    } else {
-                        1.0
+                TessDrawType::Gradient(gradient) => {
+                    let out_gradient = GradientUniforms {
+                        matrix: gradient.matrix,
+                        gradient_type: gradient.gradient_type,
+                        ratios: gradient.ratios,
+                        colors: gradient.colors,
+                        num_colors: gradient.num_colors,
+                        repeat_mode: gradient.repeat_mode,
+                        focal_point: gradient.focal_point,
                     };
-
-                    let mut options = StrokeOptions::default()
-                        .with_line_width(width)
-                        .with_line_join(match style.join_style {
-                            swf::LineJoinStyle::Round => tessellation::LineJoin::Round,
-                            swf::LineJoinStyle::Bevel => tessellation::LineJoin::Bevel,
-                            swf::LineJoinStyle::Miter(_) => tessellation::LineJoin::MiterClip,
-                        })
-                        .with_start_cap(match style.start_cap {
-                            swf::LineCapStyle::None => tessellation::LineCap::Butt,
-                            swf::LineCapStyle::Round => tessellation::LineCap::Round,
-                            swf::LineCapStyle::Square => tessellation::LineCap::Square,
-                        })
-                        .with_end_cap(match style.end_cap {
-                            swf::LineCapStyle::None => tessellation::LineCap::Butt,
-                            swf::LineCapStyle::Round => tessellation::LineCap::Round,
-                            swf::LineCapStyle::Square => tessellation::LineCap::Square,
-                        });
-
-                    if let swf::LineJoinStyle::Miter(limit) = style.join_style {
-                        options = options.with_miter_limit(limit);
-                    }
-
-                    if let Err(e) = stroke_tess.tessellate_path(
-                        &ruffle_path_to_lyon_path(commands, is_closed),
-                        &options,
-                        &mut buffers_builder,
-                    ) {
-                        // This may just be a degenerate path; skip it.
-                        log::error!("Tessellation failure: {:?}", e);
-                        continue;
+                    Draw {
+                        draw_type: DrawType::Gradient(out_gradient),
+                        vertex_buffer,
+                        index_buffer,
                     }
                 }
-            }
+                TessDrawType::Bitmap(bitmap) => {
+                    let uniforms = BitmapUniforms {
+                        matrix: bitmap.matrix,
+                        id: bitmap.id,
+                    };
+                    Draw {
+                        draw_type: DrawType::Bitmap {
+                            uniforms,
+                            is_smoothed: bitmap.is_smoothed,
+                            is_repeating: bitmap.is_repeating,
+                        },
+                        vertex_buffer,
+                        index_buffer,
+                    }
+                }
+            };
+
+            draws.push(out_draw);
         }
 
-        flush_draw(DrawType::Color, &mut mesh, &mut lyon_mesh, &self.display);
-
-        self.meshes.push(mesh);
+        self.meshes.push(Mesh { draws });
 
         handle
     }
@@ -1199,83 +942,6 @@ enum DrawType {
     },
 }
 
-fn point(x: Twips, y: Twips) -> lyon::math::Point {
-    lyon::math::Point::new(x.to_pixels() as f32, y.to_pixels() as f32)
-}
-
-fn ruffle_path_to_lyon_path(commands: Vec<DrawCommand>, is_closed: bool) -> Path {
-    let mut builder = Path::builder();
-    for cmd in commands {
-        match cmd {
-            DrawCommand::MoveTo { x, y } => {
-                builder.move_to(point(x, y));
-            }
-            DrawCommand::LineTo { x, y } => {
-                builder.line_to(point(x, y));
-            }
-            DrawCommand::CurveTo { x1, y1, x2, y2 } => {
-                builder.quadratic_bezier_to(point(x1, y1), point(x2, y2));
-            }
-        }
-    }
-
-    if is_closed {
-        builder.close();
-    }
-
-    builder.build()
-}
-
-#[allow(clippy::many_single_char_names)]
-fn swf_to_gl_matrix(m: swf::Matrix) -> [[f32; 3]; 3] {
-    let tx = m.translate_x.get() as f32;
-    let ty = m.translate_y.get() as f32;
-    let det = m.scale_x * m.scale_y - m.rotate_skew_1 * m.rotate_skew_0;
-    let mut a = m.scale_y / det;
-    let mut b = -m.rotate_skew_1 / det;
-    let mut c = -(tx * m.scale_y - m.rotate_skew_1 * ty) / det;
-    let mut d = -m.rotate_skew_0 / det;
-    let mut e = m.scale_x / det;
-    let mut f = (tx * m.rotate_skew_0 - m.scale_x * ty) / det;
-
-    a *= 20.0 / 32768.0;
-    b *= 20.0 / 32768.0;
-    d *= 20.0 / 32768.0;
-    e *= 20.0 / 32768.0;
-
-    c /= 32768.0;
-    f /= 32768.0;
-    c += 0.5;
-    f += 0.5;
-    [[a, d, 0.0], [b, e, 0.0], [c, f, 1.0]]
-}
-
-#[allow(clippy::many_single_char_names)]
-fn swf_bitmap_to_gl_matrix(m: swf::Matrix, bitmap_width: u32, bitmap_height: u32) -> [[f32; 3]; 3] {
-    let bitmap_width = bitmap_width as f32;
-    let bitmap_height = bitmap_height as f32;
-
-    let tx = m.translate_x.get() as f32;
-    let ty = m.translate_y.get() as f32;
-    let det = m.scale_x * m.scale_y - m.rotate_skew_1 * m.rotate_skew_0;
-    let mut a = m.scale_y / det;
-    let mut b = -m.rotate_skew_1 / det;
-    let mut c = -(tx * m.scale_y - m.rotate_skew_1 * ty) / det;
-    let mut d = -m.rotate_skew_0 / det;
-    let mut e = m.scale_x / det;
-    let mut f = (tx * m.rotate_skew_0 - m.scale_x * ty) / det;
-
-    a *= 20.0 / bitmap_width;
-    b *= 20.0 / bitmap_width;
-    d *= 20.0 / bitmap_height;
-    e *= 20.0 / bitmap_height;
-
-    c /= bitmap_width;
-    f /= bitmap_height;
-
-    [[a, d, 0.0], [b, e, 0.0], [c, f, 1.0]]
-}
-
 /// Returns the drawing parameters for masking.
 #[inline]
 fn mask_draw_parameters(
@@ -1344,27 +1010,5 @@ fn bitmap_blend() -> glium::Blend {
             destination: LinearBlendingFactor::OneMinusSourceAlpha,
         },
         ..Default::default()
-    }
-}
-
-struct RuffleVertexCtor {
-    color: [f32; 4],
-}
-
-impl FillVertexConstructor<Vertex> for RuffleVertexCtor {
-    fn new_vertex(&mut self, position: lyon::math::Point, _: FillAttributes) -> Vertex {
-        Vertex {
-            position: [position.x, position.y],
-            color: self.color,
-        }
-    }
-}
-
-impl StrokeVertexConstructor<Vertex> for RuffleVertexCtor {
-    fn new_vertex(&mut self, position: lyon::math::Point, _: StrokeAttributes) -> Vertex {
-        Vertex {
-            position: [position.x, position.y],
-            color: self.color,
-        }
     }
 }
